requirements.txt
dj-database-url==0.4.1
Django==1.10.1
django-crispy-forms==1.6.0
django-filter==0.15.2
djangorestframework==3.4.7
djangorestframework-word-filter==0.1.1
gunicorn==19.6.0
psycopg2==2.6.2
Pygments==2.1.3
wheel==0.24.0
whitenoise==3.2.2
